requirements.txt
PyQt6
opencv-python-headless
numpy
requests
python-socketio
